Move fetchSequence function out of lgv model
